requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.6
django-compat==1.0.15
django-cors-headers==4.3.1
django4-background-tasks==1.2.9
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
filelock==3.14.0
idna==3.7
packaging==24.0
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.1
psycopg==3.1.19
psycopg2-binary==2.9.9
PyJWT==2.8.0
requests==2.32.2
rest-framework-simplejwt==0.0.2
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
urllib3==2.2.1
virtualenv==20.26.1
